# out/v1/static/docs/api.yaml
swagger: '2.0'
info:
  title: REX
  description: Requirements Management Server API.
  version: 1.0.0
host: localhost:3000
basePath: '/v1'
schemes:
  - http
  
paths:
  '/requirement/{id}':
    get:
      tags:
        - Requirement
      summary: Get a requirement
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          required: true
          description: The requirement name or ID
          type: string
          
      responses:
        '200':
          description: Found requirement.
          schema:
            $ref: '#/definitions/RequirementSimple'
        '404':
          description: Requirement not found.
          schema:
            $ref: '#/definitions/Error'
    put:
      tags:
        - Requirement
      summary: Update or create a requirement
      consumes: 
        - application/json
      parameters:
        - in: path
          name: id
          required: true
          description: The requirement name or ID
          type: string
          
        - in: body
          name: requirement
          required: true
          schema:
              $ref: '#/definitions/RequirementFull'
              
      responses:
        '201':
          description: Requirement has been created
        '200':
          description: Requirement has been updated
        '400':
          description: Could not create or update the requirement.
          schema:
            $ref: '#/definitions/Error'
    delete:
      tags:
        - Requirement
      summary: Delete a requirement
      description: >
        Note - A requirement deletion will not remove the reqiurement from the database.
        Instead, all data is removed, and the history is updated with the removal and corresponding
        log information.
      parameters:
        - in: path
          name: id
          required: true
          description: The requirement name or ID
          type: string
          
        - in: body
          name: log
          schema:
              $ref: '#/definitions/Log'
                
      responses:
        '200':
          description: Requirement has been deleted.
        '404':
          description: Could not find the requirement.
          schema:
            $ref: '#/definitions/Error'
            
  '/requirement/history/{id}':
    get:
      tags:
        - History
      summary: Get a requirement's history
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          required: true
          description: The requirement name or ID
          type: string
      responses:
        '200':
          description: Found history.
          schema:
            $ref: '#/definitions/History'
        '404':
          description: History not found.
          schema:
            $ref: '#/definitions/Error'
  
  '/requirement/history/{id}/{version}':
    get:
      tags:
        - History
      summary: Get a specific version of a requirement
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          required: true
          description: The requirement name or ID
          type: string
        
        - in: path
          name: version
          required: true
          description: The version of the requirement
          type: integer
      responses:
        '200':
          description: Found version.
          schema:
            $ref: '#/definitions/RequirementSimple'
        '404':
          description: Version not found.
          schema:
            $ref: '#/definitions/Error'
            
  '/requirement/history/{id}/{version}/log':
    put:
      tags:
        - History
      summary: Change the log message of a specific requirement version
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          required: true
          description: The requirement name or ID
          type: string
        
        - in: path
          name: version
          required: true
          description: The requirement version
          type: string
        
        - in: body
          name: log
          required: true
          description: The updated log
          schema:
            $ref: '#/definitions/Log'
          
      responses:
        '200':
          description: Log updated
        '404':
          description: Requirement or version not found.
          schema:
            $ref: '#/definitions/Error'
        '400':
          description: Bad log body.
          schema:
            $ref: '#/definitions/Error'
    
              
definitions:
    RequirementFull:
      required: ["data"]
      properties:
        data:
          type: object
        log:
          type: string
      example:
        data:
          Description: The main loop shall update at 10hz
          Derived: true
        log: "Initial version of the requirement"
    RequirementSimple:
      required: ["data"]
      properties:
        data:
          type: object
      example:
        data:
          Description: The main loop shall update at 10hz
          Derived: true
    Error:
      type: object
      properties:
        message:
          type: object
      example:
        message: 'Error message here'
    Log:
      type: object
      properties:
        Log:
          type: string
      example:
        log: 'Initial requirement creation'
    History:
      type: array
      items:
        type: object
        properties:
          version:
            type: integer
          log:
            type: string
      example:
        - version: 0
          log: "Initial version of requirement"
        
        - version: 1
          log: "Next change"
      